backend/requirements.txt
asgiref==3.6.0
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
decorator==5.1.1
Django==4.1.6
django-cors-headers==3.13.0
django-stubs==1.14.0
django-stubs-ext==0.7.0
djangorestframework==3.14.0
djangorestframework-stubs==1.8.0
exceptiongroup==1.1.0
executing==1.2.0
factory-boy==3.2.1
Faker==17.0.0
idna==3.4
iniconfig==2.0.0
ipython==8.10.0
jedi==0.18.2
matplotlib-inline==0.1.6
mypy==1.0.0
mypy-extensions==1.0.0
packaging==23.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
prompt-toolkit==3.0.36
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pytest==7.2.1
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
types-pytz==2022.7.1.0
types-PyYAML==6.0.12.6
types-requests==2.28.11.12
types-urllib3==1.26.25.5
typing_extensions==4.4.0
urllib3==1.26.14
wcwidth==0.2.6
